requirements.txt
fire==0.1.3
GitPython==2.1.11
